fix(component): split maps component into server/client
